requirements.txt
flask
torch
transformers
pillow
gunicorn

